+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - Added the possibility to mark bell notifications as read/unread via the newly added icon button that appears on hover !4035 !4091 @anedomansky
 - Added the option to actively follow a forum, so that you get a bell for every new thread #2346 !4064 #2371 @AntonBallmaier
 - Show KAM roles on profile pages #2354 !4084 @AntonBallmaier
-- Added a dashboard container for pickup options #1279 !4096 @AntonBallmaier
+- Added a dashboard container for pickup options #1279 !4096 #2384 !4157 @AntonBallmaier
 - Add Feedback form for Beta-Testers and Bump Sentry to 9.8.0 #3968 #4113 @SaiCode
 - Added forum-like reaction to wall posts on some wall types #2355 !4079 @AntonBallmaier
 - Displayed the region of a store directly on the store page !4129 @DL6ER